+++ src/Http/Output/AbstractOutput.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Parable\Http\Output;
+
+abstract class AbstractOutput implements OutputInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function acceptsContent($content)
+    {
+        return true;
+    }
+}